Apply review comments:
* switch hostnameIsEmpty/hostnameDisallowed
* move resolveContainerTags to writer
